backend/requirements.txt
alembic==1.16.4
Flask==3.1.1
Flask_Bcrypt==1.0.1
flask_cors==6.0.1
flask_jwt_extended==4.7.1
flask_limiter==3.12
Flask_Migrate==4.1.0
flask_sqlalchemy==3.1.1
python-dotenv==1.1.1
SQLAlchemy==2.0.41
Werkzeug==3.1.3
